requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
numpy
pillow
tensorflow
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
